migrated first wave of tests over
