+++ test/functional/legacy/003_cindent_spec.lua
@@ -8,6 +8,14 @@ local execute, expect = helpers.execute, helpers.expect
 describe('cindent', function()
   setup(clear)
 
+  before_each(clear)
+
+  before_each(function ()
+      execute('set cin ts=4 sw=4')
+      execute('set nocompatible viminfo+=nviminfo modeline')
+      execute('edit                " read modeline')
+  end)
+
   it('indents blocks correctly', function()
       insert([[
       /* start of AUTO matically checked vim: set ts=4 : */
@@ -239,6 +247,648 @@ describe('cindent', function()
       	s = "/*"; b = ';'
       		s = "/*"; b = ';';
       	a = b;
+      }
+      
+      {
+      	switch (a)
+      	{
+      		case a:
+      			switch (t)
+      			{
+      				case 1:
+      					cmd;
+      					break;
+      				case 2:
+      					cmd;
+      					break;
+      			}
+      			cmd;
+      			break;
+      		case b:
+      			{
+      				int i;
+      				cmd;
+      			}
+      			break;
+      		case c: {
+      					int i;
+      					cmd;
+      				}
+      		case d: if (cond &&
+      						test) {		/* this line doesn't work right */
+      					int i;
+      					cmd;
+      				}
+      				break;
+      	}
+      }
+      
+      {
+      	if (!(vim_strchr(p_cpo, CPO_BUFOPTGLOB) != NULL && entering) &&
+      			(bp_to->b_p_initialized ||
+      			 (!entering && vim_strchr(p_cpo, CPO_BUFOPT) != NULL)))
+      		return;
+      label :
+      	asdf = asdf ?
+      		asdf : asdf;
+      	asdf = asdf ?
+      		asdf: asdf;
+      }
+      
+      /* Special Comments	: This function has the added complexity (compared  */
+      /*					: to addtolist) of having to check for a detail     */
+      /*					: texture and add that to the list first.	 	    */
+      
+      char *(array[100]) = {
+      	"testje",
+      	"foo",
+      	"bar",
+      }
+      
+      enum soppie
+      {
+      yes = 0,
+      no,
+      maybe
+      };
+      
+      typedef enum soppie
+      {
+      yes = 0,
+      no,
+      maybe
+      };
+      
+      static enum
+      {
+      yes = 0,
+      no,
+      maybe
+      } soppie;
+      
+      public static enum
+      {
+      yes = 0,
+      no,
+      maybe
+      } soppie;
+      
+      static private enum
+      {
+      yes = 0,
+      no,
+      maybe
+      } soppie;
+      
+      {
+      	int a,
+      		b;
+      }
+      
+      {
+      	struct Type
+      	{
+      		int i;
+      		char *str;
+      	} var[] =
+      	{
+      		0, "zero",
+      		1, "one",
+      		2, "two",
+      		3, "three"
+      	};
+      
+      	float matrix[3][3] =
+      	{
+      		{
+      			0,
+      			1,
+      			2
+      		},
+      		{
+      			3,
+      			4,
+      			5
+      		},
+      		{
+      			6,
+      			7,
+      			8
+      		}
+      	};
+      }
+      
+      {
+      	/* blah ( blah */
+      	/* where does this go? */
+      
+      	/* blah ( blah */
+      	cmd;
+      
+      	func(arg1,
+      			/* comment */
+      			arg2);
+      	a;
+      	{
+      		b;
+      		{
+      			c; /* Hey, NOW it indents?! */
+      		}
+      	}
+      
+      	{
+      		func(arg1,
+      				arg2,
+      				arg3);
+      		/* Hey, what am I doing here?  Is this coz of the ","? */
+      	}
+      }
+      
+      main ()
+      {
+      	if (cond)
+      	{
+      		a = b;
+      	}
+      	if (cond) {
+      		a = c;
+      	}
+      	if (cond)
+      		a = d;
+      	return;
+      }
+      
+      {
+      	case 2: if (asdf &&
+      					asdfasdf)
+      				aasdf;
+      			a = 9;
+      	case 3: if (asdf)
+      				aasdf;
+      			a = 9;
+      	case 4:    x = 1;
+      			   y = 2;
+      
+      label:	if (asdf)
+      			here;
+      
+      label:  if (asdf &&
+      				asdfasdf)
+      		{
+      		}
+      
+      label:  if (asdf &&
+      				asdfasdf) {
+      			there;
+      		}
+      
+      label:  if (asdf &&
+      				asdfasdf)
+      			there;
+      }
+      
+      {
+      	/*
+      	   hello with ":set comments= cino=c5"
+      	 */
+      
+      	/*
+      	   hello with ":set comments= cino="
+      	 */
+      }
+      
+      
+      {
+      	if (a < b) {
+      		a = a + 1;
+      	} else
+      		a = a + 2;
+      
+      	if (a)
+      		do {
+      			testing;
+      		} while (asdfasdf);
+      	a = b + 1;
+      	asdfasdf
+      }
+      
+      class bob
+      {
+      	int foo() {return 1;}
+      		int bar;
+      }
+      
+      main()
+      {
+      while(1)
+      if (foo)
+      {
+      bar;
+      }
+      else {
+      asdf;
+      }
+      misplacedline;
+      }
+      
+      {
+      	if (clipboard.state == SELECT_DONE
+      	&& ((row == clipboard.start.lnum
+      	&& col >= clipboard.start.col)
+      	|| row > clipboard.start.lnum))
+      }
+      
+      {
+      if (1) {i += 4;}
+      where_am_i;
+      return 0;
+      }
+      
+      {
+      {
+      } // sdf(asdf
+      if (asdf)
+      asd;
+      }
+      
+      {
+      label1:
+      label2:
+      }
+      
+      {
+      int fooRet = foo(pBar1, false /*fKB*/,
+      	true /*fPTB*/, 3 /*nT*/, false /*fDF*/);
+      f() {
+      for ( i = 0;
+      	i < m;
+      	/* c */ i++ ) {
+      a = b;
+      }
+      }
+      }
+      
+      {
+      	f1(/*comment*/);
+      	f2();
+      }
+      
+      {
+      do {
+      if (foo) {
+      } else
+      ;
+      } while (foo);
+      foo();	// was wrong
+      }
+      
+      int x;	    // no extra indent because of the ;
+      void func()
+      {
+      }
+      
+      char *tab[] = {"aaa",
+      	"};", /* }; */ NULL}
+      	int indented;
+      {}
+      
+      char *a[] = {"aaa", "bbb",
+      	"ccc", NULL};
+      // here
+      
+      char *tab[] = {"aaa",
+      	"xx", /* xx */};    /* asdf */
+      int not_indented;
+      
+      {
+      	do {
+      		switch (bla)
+      		{
+      			case 1: if (foo)
+      						bar;
+      		}
+      	} while (boo);
+      					wrong;
+      }
+      
+      int	foo,
+      	bar;
+      int foo;
+      
+      #if defined(foo) \
+      	&& defined(bar)
+      char * xx = "asdf\
+      	foo\
+      	bor";
+      int x;
+      
+      char    *foo = "asdf\
+      	asdf\
+      	asdf",
+      	*bar;
+      
+      void f()
+      {
+      #if defined(foo) \
+      	&& defined(bar)
+      char    *foo = "asdf\
+      	asdf\
+      	asdf",
+      	*bar;
+      	{
+      	int i;
+      char    *foo = "asdf\
+      	asdf\
+      	asdf",
+      	*bar;
+      	}
+      #endif
+      }
+      #endif
+      
+      int y;		// comment
+      		// comment
+      
+      	// comment
+      
+      {
+      	Constructor(int a,
+      			int b )  : BaseClass(a)
+      	{
+      	}
+      }
+      
+      void foo()
+      {
+      	char one,
+      	two;
+      	struct bla piet,
+      	jan;
+      	enum foo kees,
+      	jannie;
+      	static unsigned sdf,
+      	krap;
+      	unsigned int piet,
+      	jan;
+      	int
+      	kees,
+      	jan;
+      }
+      
+      {
+      	t(int f,
+      			int d);		// )
+      	d();
+      }
+      
+      Constructor::Constructor(int a,
+                               int b 
+                              )  : 
+         BaseClass(a,
+                   b,
+                   c),
+         mMember(b),
+      {
+      }
+      
+      Constructor::Constructor(int a,
+                               int b )  : 
+         BaseClass(a)
+      {
+      }
+      
+      Constructor::Constructor(int a,
+                               int b ) /*x*/ : /*x*/ BaseClass(a),
+                                                     member(b)
+      {
+      }
+      
+      class CAbc :
+         public BaseClass1,
+         protected BaseClass2
+      {
+         int Test() { return FALSE; }
+         int Test1() { return TRUE; }
+      
+         CAbc(int a, int b )  : 
+            BaseClass(a)
+         { 
+            switch(xxx)
+            {
+               case abc:
+                  asdf();
+                  break;
+      
+               case 999:
+                  baer();
+                  break;
+            }
+         }
+      
+      public: // <-- this was incoreectly indented before!!
+         void testfall();
+      protected:
+         void testfall();
+      };
+      
+      class CAbc : public BaseClass1,
+                   protected BaseClass2
+      {
+      };
+      
+      static struct
+      {
+          int a;
+          int b;
+      } variable[COUNT] =
+      {
+          {
+              123,
+              456
+          },
+      	{
+              123,
+              456
+          }
+      };
+      
+      static struct
+      {
+          int a;
+          int b;
+      } variable[COUNT] =
+      {
+          { 123, 456 },
+      	{ 123, 456 }
+      };
+      
+      void asdf()		/* ind_maxparen may cause trouble here */
+      {
+      	if ((0
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1)) break;
+      }
+      
+      foo()
+      {
+      	a = cond ? foo() : asdf
+      					   + asdf;
+      
+      	a = cond ?
+      		foo() : asdf
+      				+ asdf;
+      }
+      
+      int  main(void)
+      {
+      	if (a)
+      		if (b)
+      			2;
+      		else 3;
+      	next_line_of_code();
+      }
+      
+      barry()
+      {
+      	Foo::Foo (int one,
+      			int two)
+      		: something(4)
+      	{}
+      }
+      
+      barry()
+      {
+      	Foo::Foo (int one, int two)
+      		: something(4)
+      	{}
+      }
+      
+      Constructor::Constructor(int a,
+      		int b 
+      		)  : 
+      	BaseClass(a,
+      			b,
+      			c),
+      	mMember(b)
+      {
+      }
+             int main ()
+             {
+      	 if (lala)
+      	   do
+      	     ++(*lolo);
+      	   while (lili
+      		  && lele);
+      	   lulu;
+             }
+      
+      int main ()
+      {
+      switch (c)
+      {
+      case 'c': if (cond)
+      {
+      }
+      }
+      }
+      
+      main()
+      {
+      	(void) MyFancyFuasdfadsfnction(
+      			argument);
+      }
+      
+      main()
+      {
+      	char	foo[] = "/*";
+      	/* as
+      	df */
+      		hello
+      }
+      
+      /* valid namespaces with normal indent */
+      namespace
+      {
+      {
+        111111111111;
+      }
+      }
+      namespace /* test */
+      {
+        11111111111111111;
+      }
+      namespace // test
+      {
+        111111111111111111;
+      }
+      namespace
+      {
+        111111111111111111;
+      }
+      namespace test
+      {
+        111111111111111111;
+      }
+      namespace{
+        111111111111111111;
+      }
+      namespace test{
+        111111111111111111;
+      }
+      namespace {
+        111111111111111111;
+      }
+      namespace test {
+        111111111111111111;
+      namespace test2 {
+        22222222222222222;
+      }
+      }
+      
+      /* invalid namespaces use block indent */
+      namespace test test2 {
+        111111111111111111111;
+      }
+      namespace11111111111 {
+        111111111111;
+      }
+      namespace() {
+        1111111111111;
+      }
+      namespace()
+      {
+        111111111111111111;
+      }
+      namespace test test2
+      {
+        1111111111111111111;
+      }
+      namespace111111111
+      {
+        111111111111111111;
       }
       
       /* end of AUTO */
@@ -483,9 +1133,650 @@ describe('cindent', function()
       	a = b;
       }
       
+      {
+      	switch (a)
+      	{
+      		case a:
+      			switch (t)
+      			{
+      				case 1:
+      					cmd;
+      					break;
+      				case 2:
+      					cmd;
+      					break;
+      			}
+      			cmd;
+      			break;
+      		case b:
+      			{
+      				int i;
+      				cmd;
+      			}
+      			break;
+      		case c: {
+      					int i;
+      					cmd;
+      				}
+      		case d: if (cond &&
+      						test) {		/* this line doesn't work right */
+      					int i;
+      					cmd;
+      				}
+      				break;
+      	}
+      }
+      
+      {
+      	if (!(vim_strchr(p_cpo, CPO_BUFOPTGLOB) != NULL && entering) &&
+      			(bp_to->b_p_initialized ||
+      			 (!entering && vim_strchr(p_cpo, CPO_BUFOPT) != NULL)))
+      		return;
+      label :
+      	asdf = asdf ?
+      		asdf : asdf;
+      	asdf = asdf ?
+      		asdf: asdf;
+      }
+      
+      /* Special Comments	: This function has the added complexity (compared  */
+      /*					: to addtolist) of having to check for a detail     */
+      /*					: texture and add that to the list first.	 	    */
+      
+      char *(array[100]) = {
+      	"testje",
+      	"foo",
+      	"bar",
+      }
+      
+      enum soppie
+      {
+      	yes = 0,
+      	no,
+      	maybe
+      };
+      
+      typedef enum soppie
+      {
+      	yes = 0,
+      	no,
+      	maybe
+      };
+      
+      static enum
+      {
+      	yes = 0,
+      	no,
+      	maybe
+      } soppie;
+      
+      public static enum
+      {
+      	yes = 0,
+      	no,
+      	maybe
+      } soppie;
+      
+      static private enum
+      {
+      	yes = 0,
+      	no,
+      	maybe
+      } soppie;
+      
+      {
+      	int a,
+      		b;
+      }
+      
+      {
+      	struct Type
+      	{
+      		int i;
+      		char *str;
+      	} var[] =
+      	{
+      		0, "zero",
+      		1, "one",
+      		2, "two",
+      		3, "three"
+      	};
+      
+      	float matrix[3][3] =
+      	{
+      		{
+      			0,
+      			1,
+      			2
+      		},
+      		{
+      			3,
+      			4,
+      			5
+      		},
+      		{
+      			6,
+      			7,
+      			8
+      		}
+      	};
+      }
+      
+      {
+      	/* blah ( blah */
+      	/* where does this go? */
+      
+      	/* blah ( blah */
+      	cmd;
+      
+      	func(arg1,
+      			/* comment */
+      			arg2);
+      	a;
+      	{
+      		b;
+      		{
+      			c; /* Hey, NOW it indents?! */
+      		}
+      	}
+      
+      	{
+      		func(arg1,
+      				arg2,
+      				arg3);
+      		/* Hey, what am I doing here?  Is this coz of the ","? */
+      	}
+      }
+      
+      main ()
+      {
+      	if (cond)
+      	{
+      		a = b;
+      	}
+      	if (cond) {
+      		a = c;
+      	}
+      	if (cond)
+      		a = d;
+      	return;
+      }
+      
+      {
+      	case 2: if (asdf &&
+      					asdfasdf)
+      				aasdf;
+      			a = 9;
+      	case 3: if (asdf)
+      				aasdf;
+      			a = 9;
+      	case 4:    x = 1;
+      			   y = 2;
+      
+      label:	if (asdf)
+      			here;
+      
+      label:  if (asdf &&
+      				asdfasdf)
+      		{
+      		}
+      
+      label:  if (asdf &&
+      				asdfasdf) {
+      			there;
+      		}
+      
+      label:  if (asdf &&
+      				asdfasdf)
+      			there;
+      }
+      
+      {
+      	/*
+      	   hello with ":set comments= cino=c5"
+      	 */
+      
+      	/*
+      	   hello with ":set comments= cino="
+      	 */
+      }
+      
+      
+      {
+      	if (a < b) {
+      		a = a + 1;
+      	} else
+      		a = a + 2;
+      
+      	if (a)
+      		do {
+      			testing;
+      		} while (asdfasdf);
+      	a = b + 1;
+      	asdfasdf
+      }
+      
+      class bob
+      {
+      	int foo() {return 1;}
+      	int bar;
+      }
+      
+      main()
+      {
+      	while(1)
+      		if (foo)
+      		{
+      			bar;
+      		}
+      		else {
+      			asdf;
+      		}
+      	misplacedline;
+      }
+      
+      {
+      	if (clipboard.state == SELECT_DONE
+      			&& ((row == clipboard.start.lnum
+      					&& col >= clipboard.start.col)
+      				|| row > clipboard.start.lnum))
+      }
+      
+      {
+      	if (1) {i += 4;}
+      	where_am_i;
+      	return 0;
+      }
+      
+      {
+      	{
+      	} // sdf(asdf
+      	if (asdf)
+      		asd;
+      }
+      
+      {
+      label1:
+      label2:
+      }
+      
+      {
+      	int fooRet = foo(pBar1, false /*fKB*/,
+      			true /*fPTB*/, 3 /*nT*/, false /*fDF*/);
+      	f() {
+      		for ( i = 0;
+      				i < m;
+      				/* c */ i++ ) {
+      			a = b;
+      		}
+      	}
+      }
+      
+      {
+      	f1(/*comment*/);
+      	f2();
+      }
+      
+      {
+      	do {
+      		if (foo) {
+      		} else
+      			;
+      	} while (foo);
+      	foo();	// was wrong
+      }
+      
+      int x;	    // no extra indent because of the ;
+      void func()
+      {
+      }
+      
+      char *tab[] = {"aaa",
+      	"};", /* }; */ NULL}
+      	int indented;
+      {}
+      
+      char *a[] = {"aaa", "bbb",
+      	"ccc", NULL};
+      // here
+      
+      char *tab[] = {"aaa",
+      	"xx", /* xx */};    /* asdf */
+      int not_indented;
+      
+      {
+      	do {
+      		switch (bla)
+      		{
+      			case 1: if (foo)
+      						bar;
+      		}
+      	} while (boo);
+      	wrong;
+      }
+      
+      int	foo,
+      	bar;
+      int foo;
+      
+      #if defined(foo) \
+      	&& defined(bar)
+      char * xx = "asdf\
+      			 foo\
+      			 bor";
+      int x;
+      
+      char    *foo = "asdf\
+      				asdf\
+      				asdf",
+      		*bar;
+      
+      void f()
+      {
+      #if defined(foo) \
+      	&& defined(bar)
+      	char    *foo = "asdf\
+      					asdf\
+      					asdf",
+      			*bar;
+      	{
+      		int i;
+      		char    *foo = "asdf\
+      						asdf\
+      						asdf",
+      				*bar;
+      	}
+      #endif
+      }
+      #endif
+      
+      int y;		// comment
+      // comment
+      
+      // comment
+      
+      {
+      	Constructor(int a,
+      			int b )  : BaseClass(a)
+      	{
+      	}
+      }
+      
+      void foo()
+      {
+      	char one,
+      		 two;
+      	struct bla piet,
+      			   jan;
+      	enum foo kees,
+      			 jannie;
+      	static unsigned sdf,
+      					krap;
+      	unsigned int piet,
+      				 jan;
+      	int
+      		kees,
+      		jan;
+      }
+      
+      {
+      	t(int f,
+      			int d);		// )
+      	d();
+      }
+      
+      Constructor::Constructor(int a,
+      		int b 
+      		)  : 
+      	BaseClass(a,
+      			b,
+      			c),
+      	mMember(b),
+      {
+      }
+      
+      Constructor::Constructor(int a,
+      		int b )  : 
+      	BaseClass(a)
+      {
+      }
+      
+      Constructor::Constructor(int a,
+      		int b ) /*x*/ : /*x*/ BaseClass(a),
+      	member(b)
+      {
+      }
+      
+      class CAbc :
+      	public BaseClass1,
+      	protected BaseClass2
+      {
+      	int Test() { return FALSE; }
+      	int Test1() { return TRUE; }
+      
+      	CAbc(int a, int b )  : 
+      		BaseClass(a)
+      	{ 
+      		switch(xxx)
+      		{
+      			case abc:
+      				asdf();
+      				break;
+      
+      			case 999:
+      				baer();
+      				break;
+      		}
+      	}
+      
+      	public: // <-- this was incoreectly indented before!!
+      	void testfall();
+      	protected:
+      	void testfall();
+      };
+      
+      class CAbc : public BaseClass1,
+      	protected BaseClass2
+      {
+      };
+      
+      static struct
+      {
+      	int a;
+      	int b;
+      } variable[COUNT] =
+      {
+      	{
+      		123,
+      		456
+      	},
+      	{
+      		123,
+      		456
+      	}
+      };
+      
+      static struct
+      {
+      	int a;
+      	int b;
+      } variable[COUNT] =
+      {
+      	{ 123, 456 },
+      	{ 123, 456 }
+      };
+      
+      void asdf()		/* ind_maxparen may cause trouble here */
+      {
+      	if ((0
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1
+      				&& 1)) break;
+      }
+      
+      foo()
+      {
+      	a = cond ? foo() : asdf
+      		+ asdf;
+      
+      	a = cond ?
+      		foo() : asdf
+      		+ asdf;
+      }
+      
+      int  main(void)
+      {
+      	if (a)
+      		if (b)
+      			2;
+      		else 3;
+      	next_line_of_code();
+      }
+      
+      barry()
+      {
+      	Foo::Foo (int one,
+      			int two)
+      		: something(4)
+      	{}
+      }
+      
+      barry()
+      {
+      	Foo::Foo (int one, int two)
+      		: something(4)
+      	{}
+      }
+      
+      Constructor::Constructor(int a,
+      		int b 
+      		)  : 
+      	BaseClass(a,
+      			b,
+      			c),
+      	mMember(b)
+      {
+      }
+      int main ()
+      {
+      	if (lala)
+      		do
+      			++(*lolo);
+      		while (lili
+      				&& lele);
+      	lulu;
+      }
+      
+      int main ()
+      {
+      	switch (c)
+      	{
+      		case 'c': if (cond)
+      				  {
+      				  }
+      	}
+      }
+      
+      main()
+      {
+      	(void) MyFancyFuasdfadsfnction(
+      			argument);
+      }
+      
+      main()
+      {
+      	char	foo[] = "/*";
+      	/* as
+      	   df */
+      	hello
+      }
+      
+      /* valid namespaces with normal indent */
+      namespace
+      {
+      	{
+      		111111111111;
+      	}
+      }
+      namespace /* test */
+      {
+      	11111111111111111;
+      }
+      namespace // test
+      {
+      	111111111111111111;
+      }
+      namespace
+      {
+      	111111111111111111;
+      }
+      namespace test
+      {
+      	111111111111111111;
+      }
+      namespace{
+      	111111111111111111;
+      }
+      namespace test{
+      	111111111111111111;
+      }
+      namespace {
+      	111111111111111111;
+      }
+      namespace test {
+      	111111111111111111;
+      	namespace test2 {
+      		22222222222222222;
+      	}
+      }
+      
+      /* invalid namespaces use block indent */
+      namespace test test2 {
+      	111111111111111111111;
+      }
+      namespace11111111111 {
+      	111111111111;
+      }
+      namespace() {
+      	1111111111111;
+      }
+      namespace()
+      {
+      	111111111111111111;
+      }
+      namespace test test2
+      {
+      	1111111111111111111;
+      }
+      namespace111111111
+      {
+      	111111111111111111;
+      }
+      
       /* end of AUTO */
       ]])
 
   end)
-
 end)
